[add] first version of ExRabbitPool.Consumer macro
